+++ x-pack/plugin/ml/src/internalClusterTest/java/org/elasticsearch/xpack/ml/integration/DatafeedConfigProviderIT.java
@@ -38,11 +38,14 @@ import java.util.function.BiConsumer;
 import java.util.stream.Collectors;
 
 import static org.hamcrest.CoreMatchers.not;
+import static org.hamcrest.Matchers.allOf;
+import static org.hamcrest.Matchers.anEmptyMap;
 import static org.hamcrest.Matchers.contains;
 import static org.hamcrest.Matchers.containsInAnyOrder;
 import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.empty;
 import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.hasKey;
 import static org.hamcrest.Matchers.hasSize;
 import static org.hamcrest.Matchers.instanceOf;
 import static org.hamcrest.Matchers.nullValue;
@@ -326,7 +329,7 @@ public class DatafeedConfigProviderIT extends MlSingleNodeTestCase {
         assertThat(exceptionHolder.get().getMessage(), containsString("No datafeed with id [foo-1*] exists"));
     }
 
-    public void testFindDatafeedsForJobIds() throws Exception {
+    public void testFindDatafeedIdsForJobIds() throws Exception {
         putDatafeedConfig(createDatafeedConfig("foo-1", "j1"), Collections.emptyMap());
         putDatafeedConfig(createDatafeedConfig("foo-2", "j2"), Collections.emptyMap());
         putDatafeedConfig(createDatafeedConfig("bar-1", "j3"), Collections.emptyMap());
@@ -353,6 +356,38 @@ public class DatafeedConfigProviderIT extends MlSingleNodeTestCase {
         blockingCall(actionListener -> datafeedConfigProvider.findDatafeedIdsForJobIds(Arrays.asList("j3", "j1"), actionListener),
                 datafeedIdsHolder, exceptionHolder);
         assertThat(datafeedIdsHolder.get(), contains("bar-1", "foo-1"));
+    }
+
+    public void testFindDatafeedsForJobIds() throws Exception {
+        putDatafeedConfig(createDatafeedConfig("foo-1", "j1"), Collections.emptyMap());
+        putDatafeedConfig(createDatafeedConfig("foo-2", "j2"), Collections.emptyMap());
+        putDatafeedConfig(createDatafeedConfig("bar-1", "j3"), Collections.emptyMap());
+
+        client().admin().indices().prepareRefresh(MlConfigIndex.indexName()).get();
+
+        AtomicReference<Map<String, DatafeedConfig.Builder>> datafeedMapHolder = new AtomicReference<>();
+        AtomicReference<Exception> exceptionHolder = new AtomicReference<>();
+
+        blockingCall(
+            actionListener -> datafeedConfigProvider.findDatafeedsByJobIds(
+                Collections.singletonList("new-job"),
+                actionListener
+            ),
+            datafeedMapHolder,
+            exceptionHolder
+        );
+        assertThat(datafeedMapHolder.get(), anEmptyMap());
+
+        blockingCall(actionListener -> datafeedConfigProvider.findDatafeedsByJobIds(Collections.singletonList("j2"), actionListener),
+            datafeedMapHolder, exceptionHolder);
+        assertThat(datafeedMapHolder.get(), hasKey("j2"));
+        assertThat(datafeedMapHolder.get().get("j2").getId(), equalTo("foo-2"));
+
+        blockingCall(actionListener -> datafeedConfigProvider.findDatafeedsByJobIds(Arrays.asList("j3", "j1"), actionListener),
+            datafeedMapHolder, exceptionHolder);
+        assertThat(datafeedMapHolder.get(), allOf(hasKey("j3"), hasKey("j1")));
+        assertThat(datafeedMapHolder.get().get("j3").getId(), equalTo("bar-1"));
+        assertThat(datafeedMapHolder.get().get("j1").getId(), equalTo("foo-1"));
     }
 
     public void testHeadersAreOverwritten() throws Exception {
