+++ x-pack/plugin/ml/src/test/java/org/elasticsearch/xpack/ml/MlMetadataTests.java
@@ -26,9 +26,12 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 
 import static org.elasticsearch.xpack.core.ml.job.config.JobTests.buildJobBuilder;
 import static org.elasticsearch.xpack.ml.datafeed.DatafeedRunnerTests.createDatafeedConfig;
+import static org.hamcrest.Matchers.allOf;
+import static org.hamcrest.Matchers.hasKey;
 import static org.hamcrest.Matchers.contains;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.nullValue;
@@ -142,6 +145,22 @@ public class MlMetadataTests extends AbstractSerializingTestCase<MlMetadata> {
         assertThat(mlMetadata.expandDatafeedIds("*", false), contains("bar-1-feed", "foo-1-feed", "foo-2-feed"));
         assertThat(mlMetadata.expandDatafeedIds("foo-*", false), contains("foo-1-feed", "foo-2-feed"));
         assertThat(mlMetadata.expandDatafeedIds("foo-1-feed,bar-1*", false), contains("bar-1-feed", "foo-1-feed"));
+    }
+
+    public void testGetDatafeedsByJobIds() {
+        MlMetadata.Builder mlMetadataBuilder = newMlMetadataWithJobs("bar-1", "foo-1", "foo-2");
+        List<DatafeedConfig> datafeeds = new ArrayList<>();
+        datafeeds.add(createDatafeedConfig("bar-1-feed", "bar-1").build());
+        datafeeds.add(createDatafeedConfig("foo-1-feed", "foo-1").build());
+        datafeeds.add(createDatafeedConfig("foo-2-feed", "foo-2").build());
+        mlMetadataBuilder.putDatafeeds(datafeeds);
+        MlMetadata mlMetadata = mlMetadataBuilder.build();
+
+        Map<String, DatafeedConfig> datafeedsByJobIds = mlMetadata.getDatafeedsByJobIds(Set.of("bar-1", "foo-1", "foo-2"));
+        assertThat(datafeedsByJobIds, allOf(hasKey("bar-1"), hasKey("foo-1"), hasKey("foo-2")));
+        assertThat(datafeedsByJobIds.get("bar-1").getId(), equalTo("bar-1-feed"));
+        assertThat(datafeedsByJobIds.get("foo-1").getId(), equalTo("foo-1-feed"));
+        assertThat(datafeedsByJobIds.get("foo-2").getId(), equalTo("foo-2-feed"));
     }
 
     private static MlMetadata.Builder newMlMetadataWithJobs(String... jobIds) {
